+++ packages/smarthr-ui/src/components/ComboBox/MultiComboBox.tsx
@@ -86,7 +86,6 @@ type Props<T> = BaseProps<T> & {
     | 'aria-expanded'
     | 'aria-haspopup'
     | 'aria-invalid'
-    | 'autoComplete'
     | 'className'
     | 'onChange'
     | 'onCompositionEnd'
@@ -108,7 +107,7 @@ const multiCombobox = tv({
   slots: {
     wrapper: [
       'smarthr-ui-MultiComboBox',
-      'shr-box-border shr-inline-flex shr-min-w-[15em] shr-rounded-m shr-border shr-border-solid shr-px-0.5 shr-py-0.25',
+      'shr-box-border shr-inline-flex shr-min-w-[15em] shr-rounded-m shr-border shr-border-solid shr-px-0.5 shr-py-0.25 shr-align-bottom',
       'contrast-more:shr-border-high-contrast',
     ],
     inputArea: 'shr-flex shr-flex-1 shr-flex-wrap shr-gap-0.5 shr-overflow-y-auto',
@@ -527,7 +526,7 @@ const ActualMultiComboBox = <T,>(
             onCompositionStart={handleCompositionStartInput}
             onCompositionEnd={handleCompositionEndInput}
             onKeyDown={handleInputKeyDown}
-            autoComplete="off"
+            autoComplete={inputAttributes?.autoComplete ?? 'off'}
             tabIndex={0}
             role="combobox"
             aria-activedescendant={activeOption?.id}