+++ packages/smarthr-ui/src/types/Locale.ts
@@ -1,0 +1,1 @@
+export type LocaleMap = { [key: string]: string }